pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masq"
version = "1.0.0"
description = "Background removal (shotgun) + Lanczos upscaling engine"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Tricon Digital", email = "hello@tricondigital.com"}
]
keywords = ["image-processing", "background-removal", "upscaling", "discord-bot", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
requires-python = ">=3.10"

dependencies = [
    "Pillow>=10.0.0",
    "rembg>=2.0.50",
    "httpx>=0.25.0",
    "runware>=0.2.0",
]

[project.optional-dependencies]
bot = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
]
cli = [
    "python-dotenv>=1.0.0",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
]
all = [
    "masq[bot,cli,api]",
]

[project.scripts]
masq = "masq_cli:main"

[project.urls]
Homepage = "https://github.com/arealicehole/masq"
Documentation = "https://github.com/arealicehole/masq#readme"
Repository = "https://github.com/arealicehole/masq"
Issues = "https://github.com/arealicehole/masq/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cogs*", "masq*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]
